requirements.txt
streamlit
librosa
numpy
scipy
plotly
soundfile
pywt
mistralai
python-dotenv
